Remove Excel-based code paths
