+++ GUIA_SETUP.md
@@ -26,14 +26,13 @@ JWT_ALG=HS256
 CONFIG_PATH=./config.yaml
 CORS_ORIGINS=["http://localhost:5173"]
 UTEIS_REQ_ROOT=G:/PMA/Requerimentos Word/Modelos
-UTEIS_ALLOWED_EXTS=.pdf,.doc,.docx,.xls,.xlsx,.png,.jpg,.jpeg
+UTEIS_ALLOWED_EXTS=.pdf,.doc,.docx,.png,.jpg,.jpeg
 UTEIS_REQ_MAX_DEPTH=4
 ```
 
 Outras chaves opcionais:
 
 - `CND_DIR_BASE`, `CND_HEADLESS`, `CND_CHROME_PATH`, `CAPTCHA_MODE`, `API_KEY_2CAPTCHA` – automações Playwright de CND/CAE.
-- `PLANILHA_CERT_PATH` em produção, para apontar a planilha real de certificados/agendamentos (ver `backend/services/data_certificados.py`).
 
 ### 1.2 Banco de dados
 
@@ -66,7 +65,7 @@ python scripts/dev/mint_jwt.py --org-id 00000000-0000-0000-0000-000000000001 --s
 
 ---
 
-## 2. ETL (Excel → Postgres)
+## 2. ETL (CSV → Postgres)
 
 Utilize o mesmo `.env` do backend. Exemplos:
 
@@ -74,47 +73,18 @@ Utilize o mesmo `.env` do backend. Exemplos:
 # Ajuda geral
 python -m etl
 
-# Validar mapeamento da planilha vs config.yaml
-python -m etl debug-source caminho/planilha.xlsm
+# Validar mapeamento do arquivo vs config.yaml
+python -m etl debug-source caminho/dados.csv
 
 # Importação idempotente
-python -m etl import caminho/planilha.xlsm --dry-run   # simulação
-python -m etl import caminho/planilha.xlsm --apply     # grava no banco
+python -m etl import caminho/dados.csv --dry-run   # simulação
+python -m etl import caminho/dados.csv --apply     # grava no banco
 ```
 
-O contrato do ETL fica em `backend/etl/contracts.py` e usa as configurações de `config.yaml` para mapear abas, tabelas e aliases.
+O contrato do ETL fica em `backend/etl/contracts.py` e usa as configurações de `config.yaml` para mapear aliases e enums.
 
 ---
-
-## 3. Backend legado (opcional)
-
-Para leitura direta da planilha (sem banco), configure outro `.env` em `backend/` com os valores mínimos:
-
-```ini
-EXCEL_PATH=../data/operacional.xlsm
-CONFIG_PATH=./config.yaml
-CORS_ORIGINS=http://localhost:5173
-API_HOST=0.0.0.0
-API_PORT=8000
-LOG_LEVEL=INFO
-CND_DIR_BASE=certidoes
-CND_HEADLESS=true
-CAPTCHA_MODE=manual
-API_KEY_2CAPTCHA=
-```
-
-Execute com:
-
-```bash
-cd backend
-uvicorn backend.api:app --reload --host $API_HOST --port $API_PORT
-```
-
-As rotas `/api/*`, `/api/cnds`, `/api/cae` e `/api/certificados` ficarão disponíveis; os PDFs gerados ficam em `CND_DIR_BASE`.
-
----
-
-## 4. Frontend (React + Vite)
+## 3. Frontend (React + Vite)
 
 ```bash
 cd frontend
@@ -126,12 +96,10 @@ npm run dev
 - Para apontar para outro backend, crie `frontend/.env.local` com `VITE_API_URL=https://sua.api/econtrole`.
 - O proxy padrão do Vite redireciona `/api` para `http://localhost:8000` durante o desenvolvimento.
 
----
 
-## 5. Dicas rápidas e troubleshooting
+## 4. Dicas rápidas e troubleshooting
 
-- **Planilha não carrega?** Verifique `EXCEL_PATH`, permissões e se o arquivo não está aberto por outra pessoa (lock via `portalocker`).
-- **Coluna ausente?** Rode `python -m etl debug-source` ou `/api/diagnostico` (legado) e ajuste `backend/config.yaml`.
+- **Coluna ausente?** Rode `python -m etl debug-source` e ajuste `backend/config.yaml`.
 - **Problemas com Playwright?** Reinstale `python -m playwright install chromium` e valide dependências do Chromium no sistema.
 - **JWT inválido?** Confira `JWT_SECRET`, `JWT_ALG` e o payload usado pelo script `mint_jwt.py`.
