+++ backend/requirements.txt
@@ -9,9 +9,7 @@ SQLAlchemy==2.0.23
 alembic==1.13.1
 psycopg[binary]==3.1.12
 
-# Acesso a Excel + configuração
-openpyxl==3.1.2
-portalocker==2.8.2
+# Configuração
 pyyaml==6.0.1
 python-dotenv==1.0.0
